Custom Modification
Separate the function to the user and the admin
