+++ Front/src/components/Customization/CustomFlowerCard.jsx
@@ -1,20 +1,20 @@
-import "../../styles/Customization/Custom.css";
+// import "../../styles/Customization/Custom.css";
 
-export default function FlowerCard({ image, name, onSelect }) {
-  return (
-    <div className="card">
-      <img
-        src={image || "/images/placeholder.jpg"}
-        alt={name || "Flower"}
-        className="card-image"
-        onError={(e) => (e.target.src = "/images/placeholder.jpg")}
-      />
-      {name && <p className="card-text">{name}</p>}
-      {onSelect && (
-        <button className="select-button" onClick={onSelect}>
-          Select
-        </button>
-      )}
-    </div>
-  );
-}
+// export default function FlowerCard({ image, name, onSelect }) {
+//   return (
+//     <div className="card">
+//       <img
+//         src={image || "/images/placeholder.jpg"}
+//         alt={name || "Flower"}
+//         className="card-image"
+//         onError={(e) => (e.target.src = "/images/placeholder.jpg")}
+//       />
+//       {name && <p className="card-text">{name}</p>}
+//       {onSelect && (
+//         <button className="select-button" onClick={onSelect}>
+//           Select
+//         </button>
+//       )}
+//     </div>
+//   );
+// }
