+++ Front/src/components/Customization/CustomWrappingCard.jsx
@@ -1,13 +1,13 @@
-import "../../styles/Customization/Custom.css";
+// import "../../styles/Customization/Custom.css";
 
-export default function WrappingCard({ wrapping, onSelect }) {
-    return (
-      <div className="card">
-        <img src={wrapping.imageUrl} alt="Wrapping Paper" className="card-img" />
-        <button className="select-btn" onClick={() => onSelect(wrapping)}>
-          Select
-        </button>
-      </div>
-    );
-  }
+// export default function WrappingCard({ wrapping, onSelect }) {
+//     return (
+//       <div className="card">
+//         <img src={wrapping.imageUrl} alt="Wrapping Paper" className="card-img" />
+//         <button className="select-btn" onClick={() => onSelect(wrapping)}>
+//           Select
+//         </button>
+//       </div>
+//     );
+//   }
